Refactor stat descriptors into utility module
