+++ packages/web/src/utils/stats.ts
@@ -1,240 +1,30 @@
-import {
-	STATS,
-	POPULATION_ROLES,
-	BUILDINGS,
-	DEVELOPMENTS,
-	PHASES,
-	ACTIONS,
-	RESOURCES,
-	TRIGGER_INFO,
-	PASSIVE_INFO,
-} from '@kingdom-builder/contents';
+import { STATS, PASSIVE_INFO } from '@kingdom-builder/contents';
 import type {
 	EngineContext,
 	StatKey,
 	StatSourceContribution,
-	StatSourceLink,
 	StatSourceMeta,
 } from '@kingdom-builder/engine';
 import type { Summary, SummaryEntry } from '../translation/content/types';
+import {
+	formatDependency,
+	formatPhaseStep,
+	formatStatValue,
+	formatTriggerLabel,
+	getSourceDescriptor,
+	type SourceDescriptor,
+} from './stats/descriptors';
 
-export function statDisplaysAsPercent(key: string): boolean {
-	const info = STATS[key as keyof typeof STATS];
-	return Boolean(info?.displayAsPercent ?? info?.addFormat?.percent);
-}
-
-export function formatStatValue(key: string, value: number): string {
-	return statDisplaysAsPercent(key) ? `${value * 100}%` : String(value);
-}
+export { formatStatValue, statDisplaysAsPercent } from './stats/descriptors';
 
 function isStatKey(key: string): key is StatKey {
 	return key in STATS;
 }
 
-type SourceDescriptor = {
-	icon: string;
-	label: string;
-	suffix?: string;
-};
-
-type EntityDescriptor = {
-	icon: string;
-	label: string;
-};
-
-type DescriptorDetailFormatter = (
-	id: string | undefined,
-	detail: string | undefined,
-) => string | undefined;
-
-type DescriptorDependencyAugmenter = (
-	detail: string | undefined,
-	dep: StatSourceLink,
-	player: EngineContext['activePlayer'],
-	ctx: EngineContext,
-	options: { includeCounts?: boolean },
-) => string | undefined;
-
-type DescriptorDependencyFormatter = (
-	dep: StatSourceLink,
-	player: EngineContext['activePlayer'],
-	ctx: EngineContext,
-	options: { includeCounts?: boolean },
-) => string;
-
-type DescriptorRegistryEntry = {
-	resolve(id?: string): EntityDescriptor;
-	formatDetail?: DescriptorDetailFormatter;
-	augmentDependencyDetail?: DescriptorDependencyAugmenter;
-	formatDependency?: DescriptorDependencyFormatter;
-};
-
-const TRIGGER_LOOKUP = TRIGGER_INFO as Record<
-	string,
-	{ icon?: string; future?: string; past?: string }
->;
-
-const defaultFormatDetail: DescriptorDetailFormatter = (_id, detail) =>
-	detail ? formatDetailText(detail) : undefined;
-
-const defaultResolve =
-	(labelFallback: string): DescriptorRegistryEntry['resolve'] =>
-	(id) => ({
-		icon: '',
-		label: id ?? labelFallback,
-	});
-
-const DESCRIPTOR_REGISTRY: Record<string, DescriptorRegistryEntry> = {
-	population: {
-		resolve: (id) => {
-			const role = id
-				? POPULATION_ROLES[id as keyof typeof POPULATION_ROLES]
-				: undefined;
-			return {
-				icon: role?.icon ?? '',
-				label: role?.label ?? id ?? 'Population',
-			};
-		},
-		formatDetail: defaultFormatDetail,
-		augmentDependencyDetail: (detail, dep, player, _ctx, options) => {
-			const includeCounts = options.includeCounts ?? true;
-			if (!includeCounts || !dep.id) {
-				return detail;
-			}
-			const count = player.population?.[dep.id] ?? 0;
-			if (count > 0) {
-				return detail ? `${detail} ×${count}` : `×${count}`;
-			}
-			return detail;
-		},
-	},
-	building: {
-		resolve: (id) => {
-			if (id && BUILDINGS.has(id)) {
-				const building = BUILDINGS.get(id);
-				return { icon: building.icon ?? '', label: building.name ?? id };
-			}
-			return { icon: '', label: id ?? 'Building' };
-		},
-		formatDetail: defaultFormatDetail,
-	},
-	development: {
-		resolve: (id) => {
-			if (id && DEVELOPMENTS.has(id)) {
-				const development = DEVELOPMENTS.get(id);
-				return { icon: development.icon ?? '', label: development.name ?? id };
-			}
-			return { icon: '', label: id ?? 'Development' };
-		},
-		formatDetail: defaultFormatDetail,
-	},
-	phase: (() => {
-		const resolvePhase: DescriptorRegistryEntry['resolve'] = (id) => {
-			const phase = id ? PHASES.find((p) => p.id === id) : undefined;
-			return { icon: phase?.icon ?? '', label: phase?.label ?? id ?? 'Phase' };
-		};
-		return {
-			resolve: resolvePhase,
-			formatDetail: (id, detail) => formatStepLabel(id, detail),
-			formatDependency: (dep) => {
-				const label = formatPhaseStep(dep.id, dep.detail);
-				if (label) {
-					return label.trim();
-				}
-				const base = resolvePhase(dep.id);
-				return base.label.trim();
-			},
-		} satisfies DescriptorRegistryEntry;
-	})(),
-	action: {
-		resolve: (id) => {
-			if (id && ACTIONS.has(id)) {
-				const action = ACTIONS.get(id);
-				return { icon: action.icon ?? '', label: action.name ?? id };
-			}
-			return { icon: '', label: id ?? 'Action' };
-		},
-		formatDetail: defaultFormatDetail,
-	},
-	stat: {
-		resolve: (id) => {
-			if (id) {
-				const statInfo = STATS[id as keyof typeof STATS];
-				return { icon: statInfo?.icon ?? '', label: statInfo?.label ?? id };
-			}
-			return { icon: '', label: 'Stat' };
-		},
-		formatDetail: defaultFormatDetail,
-		augmentDependencyDetail: (detail, dep, player, ctx) => {
-			if (!dep.id) {
-				return detail;
-			}
-			const statValue =
-				player.stats?.[dep.id] ?? ctx.activePlayer.stats?.[dep.id] ?? 0;
-			const valueStr = formatStatValue(dep.id, statValue);
-			return detail ? `${detail} ${valueStr}` : valueStr;
-		},
-	},
-	resource: {
-		resolve: (id) => {
-			if (id) {
-				const resource = RESOURCES[id as keyof typeof RESOURCES];
-				return { icon: resource?.icon ?? '', label: resource?.label ?? id };
-			}
-			return { icon: '', label: 'Resource' };
-		},
-		formatDetail: defaultFormatDetail,
-	},
-	trigger: {
-		resolve: (id) => {
-			if (id) {
-				const info = TRIGGER_LOOKUP[id];
-				if (info) {
-					return {
-						icon: info.icon ?? '',
-						label: info.past ?? info.future ?? id,
-					};
-				}
-			}
-			return { icon: '', label: id ?? 'Trigger' };
-		},
-		formatDetail: defaultFormatDetail,
-	},
-	passive: {
-		resolve: () => ({
-			icon: PASSIVE_INFO.icon ?? '',
-			label: PASSIVE_INFO.label ?? 'Passive',
-		}),
-		formatDetail: defaultFormatDetail,
-	},
-	land: {
-		resolve: (id) => ({ icon: '', label: id ?? 'Land' }),
-		formatDetail: defaultFormatDetail,
-	},
-	start: {
-		resolve: () => ({ icon: '', label: 'Initial setup' }),
-		formatDetail: defaultFormatDetail,
-	},
-};
-
-function createDefaultDescriptor(kind?: string): DescriptorRegistryEntry {
-	return {
-		resolve: defaultResolve(kind ?? 'Source'),
-		formatDetail: defaultFormatDetail,
-	};
-}
-
-function getDescriptor(kind?: string): DescriptorRegistryEntry {
-	if (!kind) {
-		return createDefaultDescriptor();
-	}
-	return DESCRIPTOR_REGISTRY[kind] ?? createDefaultDescriptor(kind);
-}
-
 export function getStatBreakdownSummary(
 	statKey: string,
 	player: EngineContext['activePlayer'],
-	ctx: EngineContext,
+	context: EngineContext,
 ): Summary {
 	if (!isStatKey(statKey)) {
 		return [];
@@ -257,25 +47,8 @@ export function getStatBreakdownSummary(
 		return a.descriptor.label.localeCompare(b.descriptor.label);
 	});
 	return annotated.map(({ entry, descriptor }) =>
-		formatContribution(statKey, entry, descriptor, player, ctx),
+		formatContribution(statKey, entry, descriptor, player, context),
 	);
-}
-
-function getSourceDescriptor(meta: StatSourceMeta): SourceDescriptor {
-	const descriptorEntry = getDescriptor(meta.kind);
-	const base = descriptorEntry.resolve(meta.id);
-	const descriptor: SourceDescriptor = {
-		icon: base.icon,
-		label: base.label,
-	};
-	let suffix = descriptorEntry.formatDetail?.(meta.id, meta.detail);
-	if (suffix === undefined && meta.detail) {
-		suffix = defaultFormatDetail(meta.id, meta.detail);
-	}
-	if (suffix) {
-		descriptor.suffix = suffix;
-	}
-	return descriptor;
 }
 
 function formatContribution(
@@ -283,7 +56,7 @@ function formatContribution(
 	contribution: StatSourceContribution,
 	descriptor: SourceDescriptor,
 	player: EngineContext['activePlayer'],
-	ctx: EngineContext,
+	context: EngineContext,
 ): SummaryEntry {
 	const { amount, meta } = contribution;
 	const statInfo = STATS[statKey as keyof typeof STATS];
@@ -298,7 +71,7 @@ function formatContribution(
 		amountParts.push(statInfo.label);
 	}
 	const amountEntry = amountParts.join(' ').trim();
-	const detailEntries = buildDetailEntries(meta, player, ctx);
+	const detailEntries = buildDetailEntries(meta, player, context);
 	const title = formatSourceTitle(descriptor);
 	if (!title) {
 		if (!detailEntries.length) {
@@ -325,25 +98,27 @@ function formatSourceTitle(descriptor: SourceDescriptor): string {
 		labelParts.push(descriptor.suffix.trim());
 	}
 	if (labelParts.length) {
-		titleParts.push(
-			labelParts.length > 1
-				? `${labelParts[0]!} · ${labelParts.slice(1).join(' · ')}`
-				: labelParts[0]!,
-		);
+		const firstLabel = labelParts[0]!;
+		const remainingLabels = labelParts.slice(1).join(' · ');
+		const labelText = remainingLabels
+			? `${firstLabel} · ${remainingLabels}`
+			: firstLabel;
+		titleParts.push(labelText);
 	}
+
 	return titleParts.join(' ').trim();
 }
 
 function buildDetailEntries(
 	meta: StatSourceMeta,
 	player: EngineContext['activePlayer'],
-	ctx: EngineContext,
+	context: EngineContext,
 ): SummaryEntry[] {
 	const dependencies = (meta.dependsOn ?? [])
-		.map((dep) => formatDependency(dep, player, ctx))
+		.map((dep) => formatDependency(dep, player, context))
 		.filter((text) => text.length > 0);
 	const removal = meta.removal
-		? formatDependency(meta.removal, player, ctx, { includeCounts: false })
+		? formatDependency(meta.removal, player, context, { includeCounts: false })
 		: undefined;
 	const entries: SummaryEntry[] = [];
 	buildLongevityEntries(meta, dependencies, removal).forEach((entry) =>
@@ -447,9 +222,12 @@ function buildHistoryEntries(meta: StatSourceMeta): SummaryEntry[] {
 	if (turns.size) {
 		Array.from(turns)
 			.sort((a, b) => a - b)
-			.forEach((turn) =>
-				add(phaseHint ? `Turn ${turn} – ${phaseHint}` : `Turn ${turn}`),
-			);
+			.forEach((turn) => {
+				const label = phaseHint
+					? `Turn ${turn} – ${phaseHint}`
+					: `Turn ${turn}`;
+				add(label);
+			});
 	} else if (phaseHint) {
 		add(phaseHint);
 	}
@@ -476,25 +254,6 @@ function extractTriggerList(extra: Record<string, unknown>): string[] {
 		}
 	}
 	return triggers;
-}
-
-function formatTriggerLabel(id: string): string | undefined {
-	if (!id) {
-		return undefined;
-	}
-	const info = TRIGGER_LOOKUP[id];
-	if (info) {
-		const parts: string[] = [];
-		if (info.icon) {
-			parts.push(info.icon);
-		}
-		const label = info.past ?? info.future ?? id;
-		if (label) {
-			parts.push(label);
-		}
-		return parts.join(' ').trim();
-	}
-	return id;
 }
 
 function formatHistoryItem(entry: unknown): string | undefined {
@@ -544,110 +303,6 @@ function formatHistoryItem(entry: unknown): string | undefined {
 	return parts.join(' – ');
 }
 
-function formatPhaseStep(
-	phaseId?: string,
-	stepId?: string,
-): string | undefined {
-	if (!phaseId) {
-		return undefined;
-	}
-	const phase = PHASES.find((p) => p.id === phaseId);
-	if (!phase) {
-		return undefined;
-	}
-	const parts: string[] = [];
-	if (phase.icon) {
-		parts.push(phase.icon);
-	}
-	if (phase.label) {
-		parts.push(phase.label);
-	}
-	const base = parts.join(' ').trim();
-	const stepText = formatStepLabel(phaseId, stepId);
-	if (stepText) {
-		return base ? `${base} · ${stepText}` : stepText;
-	}
-	return base || undefined;
-}
-
-function formatStepLabel(
-	phaseId?: string,
-	stepId?: string,
-): string | undefined {
-	if (!stepId) {
-		return undefined;
-	}
-	const phase = phaseId ? PHASES.find((p) => p.id === phaseId) : undefined;
-	const step = phase?.steps.find((s) => s.id === stepId);
-	if (!step) {
-		return formatDetailText(stepId);
-	}
-	const parts: string[] = [];
-	if (step.icon) {
-		parts.push(step.icon);
-	}
-	const label = step.title ?? step.id;
-	if (label) {
-		parts.push(label);
-	}
-	return parts.join(' ').trim();
-}
-
-function formatDetailText(detail: string): string {
-	if (!detail) {
-		return '';
-	}
-	if (/^[a-z0-9]+(?:-[a-z0-9]+)*$/i.test(detail)) {
-		return detail
-			.split('-')
-			.filter((segment) => segment.length)
-			.map((segment) => segment.charAt(0).toUpperCase() + segment.slice(1))
-			.join(' ');
-	}
-	if (/^[a-z]/.test(detail)) {
-		return detail.charAt(0).toUpperCase() + detail.slice(1);
-	}
-	return detail;
-}
-
-function formatDependency(
-	dep: StatSourceLink,
-	player: EngineContext['activePlayer'],
-	ctx: EngineContext,
-	options: { includeCounts?: boolean } = {},
-): string {
-	const descriptor = getDescriptor(dep.type);
-	if (descriptor.formatDependency) {
-		return descriptor.formatDependency(dep, player, ctx, options);
-	}
-	const entity = descriptor.resolve(dep.id);
-	const fragments: string[] = [];
-	if (entity.icon) {
-		fragments.push(entity.icon);
-	}
-	if (entity.label) {
-		fragments.push(entity.label);
-	}
-	let detail = descriptor.formatDetail?.(dep.id, dep.detail);
-	if (detail === undefined && dep.detail) {
-		detail = defaultFormatDetail(dep.id, dep.detail);
-	}
-	const augmented = descriptor.augmentDependencyDetail?.(
-		detail,
-		dep,
-		player,
-		ctx,
-		options,
-	);
-	if (augmented !== undefined) {
-		detail = augmented;
-	}
-	if (detail) {
-		fragments.push(detail);
-	}
-	return fragments.join(' ').replace(/\s+/g, ' ').trim();
-}
-
 function normalizeSummaryEntry(entry: SummaryEntry): SummaryEntry | undefined {
 	if (typeof entry === 'string') {
 		const trimmed = entry.trim();
@@ -661,7 +316,11 @@ function normalizeSummaryEntry(entry: SummaryEntry): SummaryEntry | undefined {
 	if (!trimmedTitle && !normalizedItems.length) {
 		return undefined;
 	}
-	return { title: trimmedTitle || title, items: normalizedItems, ...rest };
+	return {
+		title: trimmedTitle || title,
+		items: normalizedItems,
+		...rest,
+	};
 }
 
 function pushSummaryEntry(
